--- a/src/graph/builders/explicit/dendron_note.ts
+++ b/src/graph/builders/explicit/dendron_note.ts
@@ -2,6 +2,7 @@ import { META_ALIAS } from "src/const/metadata_fields";
 import type { BCGraph } from "src/graph/MyMultiGraph";
 import type {
 	BreadcrumbsError,
+	EdgeBuilderResults,
 	ExplicitEdgeBuilder,
 } from "src/interfaces/graph";
 import type BreadcrumbsPlugin from "src/main";
@@ -48,34 +49,33 @@ const get_dendron_note_info = (
  */
 const handle_dendron_note = (
 	plugin: BreadcrumbsPlugin,
-	graph: BCGraph,
-	source_path: string,
+	results: EdgeBuilderResults,
+	source_id: string,
 	source_metadata: Record<string, unknown> | undefined,
-	errors: BreadcrumbsError[],
 ) => {
 	const { delimiter } = plugin.settings.explicit_edge_sources.dendron_note;
 
 	// NOTE: There are easier ways to do alot of the below.
 	//   But the `file` type between obsidian and dataview doesn't have the common fields needed.
 	//   So we rebuild from `path`
-	const source_basename_splits = Paths.basename(source_path).split(delimiter);
+	const source_basename_splits = Paths.basename(source_id).split(delimiter);
 	if (source_basename_splits.length === 1) return;
 
 	const dendron_note_info = get_dendron_note_info(
 		plugin,
 		source_metadata,
-		source_path,
+		source_id,
 	);
 	if (!dendron_note_info.ok) {
 		if (dendron_note_info.error) {
-			errors.push(dendron_note_info.error);
+			results.errors.push(dendron_note_info.error);
 		}
 		return;
 	}
 
-	const target_path = Paths.build(
+	const target_id = Paths.build(
 		// Use the same folder as the source
-		source_path.split("/").slice(0, -1).join("/"),
+		source_id.split("/").slice(0, -1).join("/"),
 		// Go one note up
 		source_basename_splits.slice(0, -1).join(delimiter),
 		"md",
@@ -83,57 +83,53 @@ const handle_dendron_note = (
 	const { field } = dendron_note_info.data;
 
 	// target_path is now a full path, so we can check for it directly, instead of getFirstLinkpathDest
-	const target_file = plugin.app.vault.getFileByPath(target_path);
+	const target_file = plugin.app.vault.getFileByPath(target_id);
 
 	if (!target_file) {
-		graph.safe_add_node(target_path, { resolved: false });
+		results.nodes.push({ id: target_id, attr: { resolved: false } });
 
 		// If !target_file, we can recursively call handle_dendron_note
 		//   To add the unresolved edges along the way
 		handle_dendron_note(
 			plugin,
-			graph,
-			target_path,
+			results,
+			target_id,
 			// This is really quite elegant :)
 			//   The unresolved note has no BC-dendron field, by definition
 			//   Passing undefined would just use the settings.default field
 			//   But we can propagate the field from the resolved source note
 			{ [META_ALIAS["dendron-note-field"]]: field },
-			errors,
 		);
 	}
 
-	graph.safe_add_directed_edge(source_path, target_path, {
-		field,
-		explicit: true,
-		source: "dendron_note",
+	results.edges.push({
+		source_id,
+		target_id,
+		attr: {
+			field,
+			explicit: true,
+			source: "dendron_note",
+		}
 	});
 };
 
 export const _add_explicit_edges_dendron_note: ExplicitEdgeBuilder = (
-	graph,
 	plugin,
 	all_files,
 ) => {
-	const errors: BreadcrumbsError[] = [];
+	const results: EdgeBuilderResults = { nodes: [], edges: [], errors: [] }
 
 	if (!plugin.settings.explicit_edge_sources.dendron_note.enabled) {
-		return { errors };
+		return results
 	}
 
 	all_files.obsidian?.forEach(({ file, cache }) => {
-		handle_dendron_note(
-			plugin,
-			graph,
-			file.path,
-			cache?.frontmatter,
-			errors,
-		);
+		handle_dendron_note(plugin, results, file.path, cache?.frontmatter);
 	});
 
 	all_files.dataview?.forEach((page) => {
-		handle_dendron_note(plugin, graph, page.file.path, page, errors);
+		handle_dendron_note(plugin, results, page.file.path, page);
 	});
 
-	return { errors };
+	return results
 };