--- a/src/graph/builders/explicit/johnny_decimal_note.ts
+++ b/src/graph/builders/explicit/johnny_decimal_note.ts
@@ -2,6 +2,7 @@ import { META_ALIAS } from "src/const/metadata_fields";
 import type { BCGraph } from "src/graph/MyMultiGraph";
 import type {
 	BreadcrumbsError,
+	EdgeBuilderResults,
 	ExplicitEdgeBuilder,
 } from "src/interfaces/graph";
 import type BreadcrumbsPlugin from "src/main";
@@ -48,10 +49,9 @@ const get_johnny_decimal_note_info = (
  */
 const handle_johnny_decimal_note = (
 	plugin: BreadcrumbsPlugin,
-	graph: BCGraph,
+	results: EdgeBuilderResults,
 	source_note: JohnnyDecimalNote,
 	notes: JohnnyDecimalNote[],
-	errors: BreadcrumbsError[],
 ) => {
 	const johnny_decimal_note_info = get_johnny_decimal_note_info(
 		plugin,
@@ -60,7 +60,7 @@ const handle_johnny_decimal_note = (
 	);
 	if (!johnny_decimal_note_info.ok) {
 		if (johnny_decimal_note_info.error) {
-			errors.push(johnny_decimal_note_info.error);
+			results.errors.push(johnny_decimal_note_info.error);
 		}
 		return;
 	}
@@ -85,15 +85,19 @@ const handle_johnny_decimal_note = (
 
 	// NOTE: I don't think this can ever happen... if target_note, then target_file must exist
 	if (!target_file) {
-		graph.safe_add_node(target_note.path, { resolved: false });
+		results.nodes.push({ id: target_note.path, attr: { resolved: false } });
 	}
 
 	const { field } = johnny_decimal_note_info.data;
 
-	graph.safe_add_directed_edge(source_note.path, target_note.path, {
-		field,
-		explicit: true,
-		source: "johnny_decimal_note",
+	results.edges.push({
+		source_id: source_note.path,
+		target_id: target_note.path,
+		attr: {
+			field,
+			explicit: true,
+			source: "johnny_decimal_note",
+		}
 	});
 };
 
@@ -105,14 +109,13 @@ type JohnnyDecimalNote = {
 };
 
 export const _add_explicit_edges_johnny_decimal_note: ExplicitEdgeBuilder = (
-	graph,
 	plugin,
 	all_files,
 ) => {
-	const errors: BreadcrumbsError[] = [];
+	const results: EdgeBuilderResults = { nodes: [], edges: [], errors: [] }
 
 	if (!plugin.settings.explicit_edge_sources.johnny_decimal_note.enabled) {
-		return { errors };
+		return results
 	}
 
 	const { delimiter } =
@@ -155,12 +158,11 @@ export const _add_explicit_edges_johnny_decimal_note: ExplicitEdgeBuilder = (
 	johnny_decimal_notes.forEach((note) => {
 		handle_johnny_decimal_note(
 			plugin,
-			graph,
+			results,
 			note,
 			johnny_decimal_notes,
-			errors,
 		);
 	});
 
-	return { errors };
+	return results
 };