--- a/src/graph/builders/explicit/typed_link.ts
+++ b/src/graph/builders/explicit/typed_link.ts
@@ -1,5 +1,5 @@
 import type {
-	BreadcrumbsError,
+	EdgeBuilderResults,
 	ExplicitEdgeBuilder,
 } from "src/interfaces/graph";
 import { ensure_is_array } from "src/utils/arrays";
@@ -8,11 +8,10 @@ import { resolve_relative_target_path } from "src/utils/obsidian";
 const MARKDOWN_LINK_REGEX = /\[(.+?)\]\((.+?)\)/;
 
 export const _add_explicit_edges_typed_link: ExplicitEdgeBuilder = (
-	graph,
 	plugin,
 	all_files,
 ) => {
-	const errors: BreadcrumbsError[] = [];
+	const results: EdgeBuilderResults = { nodes: [], edges: [], errors: [] }
 
 	const field_labels = new Set(
 		plugin.settings.edge_fields.map((f) => f.label),
@@ -29,7 +28,7 @@ export const _add_explicit_edges_typed_link: ExplicitEdgeBuilder = (
 				const field = target_link.key.split(".")[0];
 				if (!field_labels.has(field)) return;
 
-				const [target_path, target_file] = resolve_relative_target_path(
+				const [target_id, target_file] = resolve_relative_target_path(
 					plugin.app,
 					target_link.link,
 					source_file.path,
@@ -37,13 +36,17 @@ export const _add_explicit_edges_typed_link: ExplicitEdgeBuilder = (
 
 				if (!target_file) {
 					// Unresolved nodes don't have aliases
-					graph.safe_add_node(target_path, { resolved: false });
+					results.nodes.push({ id: target_id, attr: { resolved: false } });
 				}
 
-				graph.safe_add_directed_edge(source_file.path, target_path, {
-					field,
-					explicit: true,
-					source: "typed_link",
+				results.edges.push({
+					target_id,
+					source_id: source_file.path,
+					attr: {
+						field,
+						explicit: true,
+						source: "typed_link",
+					}
 				});
 			});
 		},
@@ -84,14 +87,14 @@ export const _add_explicit_edges_typed_link: ExplicitEdgeBuilder = (
 						// @ts-expect-error: instanceof didn't work here?
 						target_link?.isLuxonDateTime
 					) {
-						errors.push({
+						results.errors.push({
 							path: source_file.path,
 							code: "invalid_field_value",
 							message: `Invalid value for field '${field}': '${target_link}'. Dataview DateTime values are not supported, since they don't preserve the original date string.`,
 						});
 					} else {
 						// It's a BC field, with a definitely invalid value, cause it's not a link
-						errors.push({
+						results.errors.push({
 							path: source_file.path,
 							code: "invalid_field_value",
 							message: `Invalid value for field '${field}': '${target_link}'. Expected wikilink or markdown link.`,
@@ -110,22 +113,23 @@ export const _add_explicit_edges_typed_link: ExplicitEdgeBuilder = (
 					if (!target_file) {
 						// It's an unresolved link, so we add a node for it
 						// But still do it safely, as a previous file may point to the same unresolved node
-						graph.safe_add_node(target_path, { resolved: false });
+						results.nodes.push({ id: target_path, attr: { resolved: false } });
 					}
 
 					// If the file exists, we should have already added a node for it in the simple loop over all markdown files
-					graph.safe_add_directed_edge(
-						source_file.path,
-						target_path,
-						{
+					results.edges.push({
+						source_id: source_file.path,
+						target_id: target_path,
+						attr: {
 							field,
 							explicit: true,
 							source: "typed_link",
-						},
+						}
+					},
 					);
 				});
 		});
 	});
 
-	return { errors };
+	return results
 };