@mProjectsCode 2024-04-27
TODO list:

- [ ] make edge builder build the rust data directly
- [ ] relevant settings get passed to rust
- [ ] implement all features for mermaid code blocks
- [ ] implement all features for tree view
- [ ] fix all of the other views
- [ ] get incremental graph rebuilds on vault events working