@mProjectsCode 2024-05-02
Some preliminary performance metrics. (Take with a grain of salt, as with all benchmarks.)
Tested with `2564` notes in a tree shape. Each note has 1 to 3 explicit `up` frontmatter links. The tree is cut off at depth 10.
Transitive rules:
```
[up] <- down
[down] <- up
[same] <- same
[uncle] <- nephew
[nephew] <- uncle
[up, down] -> same
[up, same] -> uncle
```

This fork: 
- Initial Graph: `~4ms`
- Implied Edges: `~2.5ms`
- Total: `~18ms`

Current Master (4.2.27) after running a few times to get JIT advantages:
- Implied Edges: `~70ms`
- Total: `~85ms`

This fork is about `4x` faster regarding total time and about `25x` faster in building implied edges.