@mProjectsCode 2024-05-03
There might be some strangeness with sorting, as bind gen frees arrays of rust objects when passed to rust. Since there is no way to pass a reference to an array, the sorting code returns an array that has to override the array passed in.