@mProjectsCode 2025-01-03
Did some optimizations:
1. removed `dyn` in hot paths and replaced it with enums
2. replaced `String` with `Rc<str>` since we have lots of the same small strings, we can use `Rc`, which only clones the pointer and not the string.