@mProjectsCode 2025-01-03
I limited the tree view and trail view to depth 5 for now. The graph engine can easily handle more, but the rendering is too slow. There should be an indicator that a cut-off happened though.