@mProjectsCode 2025-01-05
The total graph rebuilding time is now at `~15ms` for a tree structure vault with 2.5k files.