@mProjectsCode 2025-03-13
One thing I would still like to do is look into better methods of limiting the number of items shown in views. Thanks to the nature of graphs we can get exponential blowup in regards to the depth there. This can majorly tank performance and responsiveness of the app due to too many elements being on the screen. The most promising option to me atm seems to be a hard limit 100 to 200 items that are shown.