@mProjectsCode 2025-06-01
Implemented the traversal hard limit and set it to 100 for all views for now.